requirements.txt
# Victor Trading System Dependencies

# Web Crawling
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# NLP / Keyword Extraction
konlpy>=0.6.0
keybert>=0.8.0
krwordrank>=1.0.3
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0

# Korea Investment API
python-kis>=1.0.0

# Slack
slack-sdk>=3.23.0

# Scheduling
APScheduler>=3.10.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Logging
loguru>=0.7.0

# Caching
diskcache>=5.6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
